if y switch con javascript

+++ 08Javascript/03_calculadora_funciones/script.js
@@ -2,8 +2,8 @@ function mostrarDatos(e) {
     e.preventDefault();
     var uno = document.querySelector("#uno").value;
     var dos = document.querySelector("#dos").value;
-    var resultado = parseInt(uno) + parseInt(dos);
-    var mensaje = uno+" + "+ dos +" = "+ resultado;	
+    var resultado = parseFloat(uno) + parseFloat(dos);
+    var mensaje = uno + " + " + dos + " = " + resultado;
     document.querySelector("h1").textContent = mensaje;
     console.log(mensaje);
 }
